refactor(trusty-memory): split lib.rs into events + http_server modules (issue #607)
Extract two coherent groups from lib.rs to bring it from 616 to 468 SLOC:

1. `events.rs` (139 SLOC): HookType, InjectionKind, DaemonEvent + impls,
   and open_activity_log_with_fallback — the event-bus vocabulary used by
   all mutation handlers, the activity log, and the SSE stream.

2. `http_server.rs` (156 SLOC): write_http_addr_file, dotfile_http_addr_path,
   run_http_on, run_http, run_http_dynamic, spawn_disk_size_ticker,
   spawn_status_event_ticker, sse_handler — all gated on `axum-server`
   feature and self-contained relative to AppState.

Public API is preserved via re-exports in lib.rs. lib_tests.rs continues
to access all items through `use super::*`. Removes the lib.rs entry from
.line-cap-allowlist.tsv (ratchet down: 616 → 468 SLOC).
